% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_vis.R
\name{SigVis}
\alias{SigVis}
\title{SIG Visualization}
\description{
The `SigVis` class provides methods to visualize SIG networks with
customizable plotting options using ggraph and ggplot2.
}
\details{
An R6 class for creating visualizations of structured intelligence 
governance networks using ggraph.


This class takes SigGraph objects and creates publication-ready
network visualizations with appropriate theming and layout.
}
\examples{
\dontrun{
# Create visualization from graph
sig_dat <- SigDat$new(source = "template")
sig_graph <- SigGraph$new(sig_dat = sig_dat)
sig_vis <- SigVis$new(sig_graph = sig_graph)

# Create plot
plot <- sig_vis$plot()
print(plot)
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sig_graph}}{SigGraph object (currently unused)}

\item{\code{graph}}{tidygraph object for visualization
Initialize SigVis object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SigVis-new}{\code{SigVis$new()}}
\item \href{#method-SigVis-get_graph}{\code{SigVis$get_graph()}}
\item \href{#method-SigVis-plot}{\code{SigVis$plot()}}
\item \href{#method-SigVis-clone}{\code{SigVis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SigVis-new"></a>}}
\if{latex}{\out{\hypertarget{method-SigVis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigVis$new(source = "template", sig_graph = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Character. Data source identifier (unused)}

\item{\code{sig_graph}}{SigGraph object containing the tidygraph to visualize}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SigVis` object
Get the graph object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SigVis-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-SigVis-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigVis$get_graph()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tidygraph object
Create network visualization plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SigVis-plot"></a>}}
\if{latex}{\out{\hypertarget{method-SigVis-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigVis$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ggplot object with network visualization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SigVis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SigVis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigVis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
