% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_nodes}
\alias{template_nodes}
\title{Template nodes data for SIG visualization}
\format{
A data frame with 8 rows and 2 variables:
\describe{
  \item{node}{Character. Name of the node}
  \item{node_context}{Character. Contextual category of the node}
}
}
\source{
Template data for demonstration and testing purposes
}
\usage{
template_nodes
}
\description{
A dataset containing example node information for structured intelligence
governance networks. This data represents different components and their
contextual classifications in a governance system.
}
\keyword{datasets}
